# back/config/swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Groupomania API",
    "description": "Groupomania social media API documentation",
    "contact": {
      "email": "nssmhttb@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
...
  "paths": {
...
    "/posts/": {
      "get": {
        "tags": [
          "post"
        ],
        "description": "This can only be done by the logged in user. Add xsrf token in **Authorize**",
        "summary": "Fetch all posts",
        "operationId": "getAllPosts",
        "security": [
          {
            "accessToken": [],
            "xsrfToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "successfully operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllPostsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "post"
        ],
        "description": "This can only be done by the logged in user. Add xsrf token in **Authorize**",
        "summary": "Create new post",
        "operationId": "createPost",
        "security": [
          {
            "accessToken": [],
            "xsrfToken": []
          }
        ],
        "requestBody": {
          "description": "Update post object",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePost"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successfully operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPostByIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/posts/{id}": {
      "parameters": [
        {
          "name": "id",
          "description": "id post",
          "required": true,
          "in": "path",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "tags": [
          "post"
        ],
        "description": "This can only be done by the logged in user. Add xsrf token in **Authorize**",
        "summary": "Fetch post by id",
        "operationId": "getPostById",
        "security": [
          {
            "accessToken": [],
            "xsrfToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "successfully operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPostByIdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "post"
        ],
        "description": "This can only be done by the logged in user and the post is owned by the user. Add xsrf token in **Authorize**",
        "summary": "Update post by id",
        "operationId": "updatePost",
        "security": [
          {
            "accessToken": [],
            "xsrfToken": []
          }
        ],
        "requestBody": {
          "description": "Update post object",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePost"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successfully operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "post"
        ],
        "description": "This can only be done by the logged in user and the post is owned by the user. Add xsrf token in **Authorize**",
        "summary": "Delete post by id",
        "operationId": "deletePost",
        "security": [
          {
            "accessToken": [],
            "xsrfToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "successfully operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/posts/like/{id}": {
      "post": {
        "tags": [
          "post"
        ],
        "security": [
          {
            "accessToken": [],
            "xsrfToken": []
          }
        ],
        "description": "This can only be done by the logged in user. Add xsrf token in **Authorize**",
        "summary": "Like post by id",
        "operationId": "likePost",
        "parameters": [
          {
            "name": "id",
            "description": "id post",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successfully operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400":{
            "description": "Bad Request",
            "content": {
              "application/json":{
                "schema":{
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HttpErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "validationError": {
            "type": "object",
            "properties": {
              "propertyName": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string"
                  },
                  "context": {
                    "type": "array",
                    "items": {
                      "minItems": 1,
                      "type": "object",
                      "required": [
                        "label",
                        "value",
                        "invalids",
                        "key"
                      ],
                      "properties": {
                        "label": {
                          "type": "string"
                        },
                        "value": {
                          "type": "string"
                        },
                        "invalids": {
                          "type": "array",
                          "items": {
                            "minItems": 1,
                            "required": [
                              "invalid"
                            ]
                          },
                          "properties": {
                            "invalid": {
                              "type": "string"
                            }
                          }
                        },
                        "key": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "type": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "GetAllPostsResponse": {
        "type": "array",
        "items": {
          "minItems": 1,
          "type": "object",
          "required": [
            "id",
            "content",
            "media",
            "created_at",
            "updated_at",
            "comment_count",
            "User"
          ],
          "properties": {
            "id": {
              "type": "string"
            },
            "content": {
              "type": "string"
            },
            "media": {
              "type": "string"
            },
            "created_at": {
              "type": "string",
              "format": "date"
            },
            "updated_at": {
              "type": "string",
              "format": "date"
            },
            "comment_count": {
              "type": "integer"
            },
            "User": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
      "GetPostByIdResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date"
          },
          "updated_at": {
            "type": "string",
            "format": "date"
          },
          "User": {
            "$ref": "#/components/schemas/User"
          },
          "Comment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Comment"
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "profilePicture": {
            "type": "string"
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          },
          "media": {
            "type": "string",
            "format": "binary"
          },
          "mediaType":{
            "type": "string"
          },
          "like":{
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date"
          },
          "updated_at": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "CreatePost": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "media": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "User": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "HttpMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "accessToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      },
      "refreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refrsh_token"
      },
      "xsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-xsrf-token"
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}